chore: refactor tests after changes in args
